WIP: FIRST WORKING VERSION
In order to save me from some _current problems_:
- BLEEDING EDGE: Already using v0.23.0-rc.1, because it changed the `buf`
  manifest/generated code to the new format (local/remote,
  `buf/protobuf/wire`);
- HACK: pinned the `protoc-gen-ts_proto` dependencies using a patched
  `package.json` (to remove the `yarn` references) + a manually
  generated `pnpm-lock.yaml` (which probably does not have the exact
  same dependencies). Yarn v2 locks aren't currently supported by
  nixpkgs, although there's work in progress in pull_355053 (avoiding
  referecing the issue because of GitHub's spam).

The overall structure I'm using to achieve the build is probably not
great from a cache/hashing updates standpoint; possibly worth
re-working.

But it does work.

FIXME: should I `substituteInPlace` the `/var/opt/memos` reference by
the more nixpkgs-standard `/var/lib/memos`? This is overridable via
flags and environment, but this is one of those hardcoded paths that
could be a bit more plug-and-play, I guess.